+++ FunctionsLib.h
@@ -3,7 +3,7 @@ namespace Elevator{
 	class Functions{
 		public:
 			// Returns a + b
-			static double TimeToDestiny(int origin_floor, int destiny_floor, double time_sended, double time_now);
-			static int ChooseController(int est1, int est2, int of1, int of2, int ff1, int ff2, double t1, double t2, int floor, double t);
+			static double TimeToDestiny(double origin_floor, double destiny_floor, double time_sended, double time_now);
+			static double ChooseController(double est1, double est2, double of1, double of2, double ff1, double ff2, double t1, double t2, double floor, double t);
   };
 }
