+++ controller.h
@@ -13,12 +13,12 @@ class controller: public Simulator {
 // Declare the state,
 // output variables
 // and parameters
-  int current_floor;
-  int final_floor;
+  double current_floor;
+  double final_floor;
   double sigma;
   double inf;
-  int output;
-  int out_value;
+  double output;
+  double out_value;
 public:
 	controller(const char *n): Simulator(n) {};
 	void init(double, ...);
